+++ lib/spandx.rb
@@ -6,6 +6,7 @@ require 'json'
 require 'net/hippie'
 require 'nokogiri'
 require 'pathname'
+require 'text'
 
 require 'spandx/catalogue'
 require 'spandx/content'