+++ lib/spandx/guess.rb
@@ -27,19 +27,42 @@ module Spandx
       @catalogue = catalogue
     end
 
-    def license_for(raw_content)
+    def license_for(raw_content, algorithm: :dice_coefficient)
       content = Content.new(raw_content)
+      score = nil
 
-      max_score = nil
-      catalogue.each do |license|
-        next if license.deprecated_license_id?
+      if algorithm == :dice_coefficient
+        catalogue.each do |license|
+          next if license.deprecated_license_id?
 
-        percentage = content.similarity_score(license.content)
-        if (percentage > 89.0) && (max_score.nil? || percentage > max_score.score)
-          max_score = Score.new(percentage, license)
+          score = dice(content, license, score)
+        end
+      elsif algorithm == :levenshtein
+        catalogue.each do |license|
+          next if license.deprecated_license_id?
+
+          score = levenshtein(content, license, score)
         end
       end
-      max_score ? max_score.item.id : nil
+      score ? score.item.id : nil
+    end
+
+    private
+
+    def levenshtein(target, other, score)
+      percentage = target.similarity_score(other.content, algorithm: :levenshtein)
+      if (score.nil? || percentage < score.score)
+        return Score.new(percentage, other)
+      end
+      score
+    end
+
+    def dice(target, other, score)
+      percentage = target.similarity_score(other.content, algorithm: :dice_coefficient)
+      if (percentage > 89.0) && (score.nil? || percentage > score.score)
+        return Score.new(percentage, other)
+      end
+      score
     end
   end
 end