+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@
 require 'bundler/setup'
 require 'spandx'
 require 'spandx/cli'
+require 'rspec-benchmark'
 require 'webmock/rspec'
 require 'securerandom'
 Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
@@ -10,6 +11,7 @@ Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'
+  config.include RSpec::Benchmark::Matchers
   config.include(Module.new do
     def fixture_file(file)
       File.join(File.dirname(__FILE__), 'fixtures', file)