+++ spec/unit/guess_spec.rb
@@ -108,9 +108,28 @@ RSpec.describe Spandx::Guess do
         pending { expect(subject.license_for(license_file(license))).to eql(license) }
       end
     end
-  end
 
-  it 'guesses the spandx license' do
-    expect(subject.license_for(IO.read('LICENSE.txt'))).to eql('MIT')
+    context "when guessing the spandx license" do
+      let(:content) { IO.read('LICENSE.txt') }
+
+      it 'guesses the spandx license using the default algorithm' do
+        expect(subject.license_for(content)).to eql('MIT')
+      end
+
+      [
+        :dice_coefficient,
+        :levenshtein
+      ].each do |algorithm|
+        context algorithm.to_s do
+          specify { expect(subject.license_for(content, algorithm: algorithm)).to eql('MIT') }
+
+          specify do
+            expect do
+              subject.license_for(content, algorithm: algorithm)
+            end.to perform_under(0.05).sample(10)
+          end
+        end
+      end
+    end
   end
 end