+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Parse .NET `sln` files
+- Add ability to choose Levenshtein algorithm
 
 ## [0.1.7] - 2020-01-28
 ### Added